backend/requirements.txt
fastapi
uvicorn[standard]
python-multipart
xgboost
numpy
pydantic-ai-slim[google]

